+++ run.py
@@ -1335,22 +1335,24 @@ def create_keystore():
     Entry(create_keystore_window, textvariable=your_two_letter_country_code, width=35).pack()
 
     def save_keystore():
-        file = filedialog.asksaveasfile(mode='w', defaultextension=".keystore")
+        if alias_name.get() == '':
+            messagebox.showerror("Error", "Please enter alias name.")
+            return False
+        elif keystore_password.get() == '':
+            messagebox.showerror("Error", "Please enter keystore password.")
+            return False
+        elif your_name.get() == '':
+            messagebox.showerror("Error", "Please enter your name.")
+            return False
+        else:
+            pass
+
+        file = filedialog.asksaveasfile(mode='w')
         if file is None:
             return
         else:
-            keystore_file =  'keytool -genkey -noprompt \
-            -alias ' + alias_name.get() + ' \
-            -keypass ' + keystore_password.get() + ' \
-            -keyalg ' + 'RSA' + ' \
-            -sigalg ' + 'SHA1withRSA' + '\
-            -validity ' + '10000' + ' \
-            -keysize ' + '2048' + ' \
-            -keystore ' + file.name + ' \
-            -storepass '+ keystore_password.get() +' \
-            -storetype  ' + 'JKS' 
-            
-            os.system(keystore_file)
+            os.system(f'keytool -genkey -noprompt -v -keystore {file.name}.keystore -keyalg RSA -keysize 2048 -validity 10000 -alias {alias_name.get()} -storetype PKCS12 -storepass {keystore_password.get()} -keypass {keystore_password.get()} -dname "CN={your_name.get()}, OU={your_organization_unit.get()}, O={your_organization.get()}, L={your_city_or_locality.get()}, S={your_state_or_province.get()}, C={your_two_letter_country_code.get()}"')
+            os.remove(file.name)
             print('Keystore created successfully.')
             create_keystore_window.destroy()
             messagebox.showinfo("Success", "Keystore created successfully.")